+++ KitchenDelivery/App/BusinesLogic/Network/NetworkService.swift
@@ -9,16 +9,16 @@ import Foundation
 import Combine
 
 protocol NetworkServiceProtocol {
-    func getCategories(url: URL) -> AnyPublisher<Category.CategoriesResponse, NetworkError>
-    func getDishes(url: URL) -> AnyPublisher<Dish.DishesResponse, NetworkError>
+    func getCategories(url: URL) -> AnyPublisher<Category.CategoriesResponse, NetworkServiceError>
+    func getDishes(url: URL) -> AnyPublisher<Dish.DishesResponse, NetworkServiceError>
 }
 
 class NetworkService: NetworkServiceProtocol {
-    func getCategories(url: URL) -> AnyPublisher<Category.CategoriesResponse, NetworkError> {
+    func getCategories(url: URL) -> AnyPublisher<Category.CategoriesResponse, NetworkServiceError> {
         URLSession.shared.publisher(for: url, queue: "CategoriesRequest")
     }
     
-    func getDishes(url: URL) -> AnyPublisher<Dish.DishesResponse, NetworkError> {
+    func getDishes(url: URL) -> AnyPublisher<Dish.DishesResponse, NetworkServiceError> {
         URLSession.shared.publisher(for: url, queue: "DishesRequest")
     }
 }
