+++ KitchenDelivery/App/BusinesLogic/Network/URLSessionExtension.swift
@@ -12,27 +12,27 @@ extension URLSession {
     func publisher<T: Decodable>(for url: URL,
                                  queue label: String,
                                  responseType: T.Type = T.self,
-                                 decoder: JSONDecoder = .init()) -> AnyPublisher<T, NetworkError> {
+                                 decoder: JSONDecoder = .init()) -> AnyPublisher<T, NetworkServiceError> {
         dataTaskPublisher(for: url)
             .receive(on: DispatchQueue(label: label, qos: .background, attributes: .concurrent))
             .map(\.data)
             .decode(type: NetworkResponse<T>.self, decoder: decoder)
-            .mapError({ error -> NetworkError in
+            .mapError({ error -> NetworkServiceError in
                 switch error {
                 case is URLError:
-                    return NetworkError.network(description: error.localizedDescription)
+                    return NetworkServiceError.network(description: error.localizedDescription)
                 case is DecodingError:
-                    return NetworkError.decodingError(description: error.localizedDescription)
+                    return NetworkServiceError.decodingError(description: error.localizedDescription)
                 default:
-                    return NetworkError.invalidResponse
+                    return NetworkServiceError.invalidResponse
                 }
             })
-            .flatMap({ response -> AnyPublisher<T, NetworkError> in
+            .flatMap({ response -> AnyPublisher<T, NetworkServiceError> in
                 guard let value = response.wrappedValue else {
-                    return Fail<T, NetworkError>(error: NetworkError.unwrap).eraseToAnyPublisher()
+                    return Fail<T, NetworkServiceError>(error: NetworkServiceError.unwrap).eraseToAnyPublisher()
                 }
                 return Just(value)
-                    .setFailureType(to: NetworkError.self)
+                    .setFailureType(to: NetworkServiceError.self)
                     .eraseToAnyPublisher()
             })
             .eraseToAnyPublisher()
