+++ KitchenDelivery/App/Flow/ContentView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct ContentView: View {
+    
     var body: some View {
         MainTabView()
     }
