+++ KitchenDelivery/App/Flow/Main/Category/View/Subviews/DishItem/ViewModel/DishItemViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  DishItemViewModel.swift
+//  KitchenDelivery
+//
+//  Created by Denis Dmitriev on 12.07.2023.
+//
+
+import Foundation
+
+class DishItemViewModel: ObservableObject {
+    
+    let dish: Dish
+    
+    init(dish: Dish) {
+        self.dish = dish
+    }
+}
