+++ KitchenDelivery/App/Flow/Main/Product/ViewModel/ProductViewModel.swift
@@ -1,0 +1,20 @@
+//
+//  ProductViewModel.swift
+//  KitchenDelivery
+//
+//  Created by Denis Dmitriev on 12.07.2023.
+//
+
+import Foundation
+import SwiftUI
+
+class ProductViewModel: ObservableObject {
+    
+    
+    let dish: Dish
+    
+    init(dish: Dish) {
+        self.dish = dish
+    }
+    
+}
