+++ KitchenDelivery/Resources/Assets.xcassets/Icons/like.fill.imageset/like.fill.pdf
@@ -1,0 +1,158 @@
+%PDF-1.7
+
+1 0 obj
+  << >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 2.666687 2.455963 cm
+0.000000 0.000000 0.000000 scn
+9.333334 14.381824 m
+7.259260 19.088074 0.000000 18.586815 0.000000 12.571690 c
+0.000000 6.556563 9.333334 1.544038 9.333334 1.544038 c
+9.333334 1.544038 18.666668 6.556563 18.666668 12.571690 c
+18.666668 18.586815 11.407409 19.088074 9.333334 14.381824 c
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 2.666687 1.669983 cm
+0.000000 0.000000 0.000000 scn
+9.333334 15.167805 m
+8.418258 14.764525 l
+8.578086 14.401859 8.937012 14.167805 9.333334 14.167805 c
+9.729656 14.167805 10.088581 14.401859 10.248410 14.764524 c
+9.333334 15.167805 l
+h
+9.333334 2.330018 m
+8.860194 1.449030 l
+9.155663 1.290348 9.511004 1.290348 9.806474 1.449030 c
+9.333334 2.330018 l
+h
+1.000000 13.357670 m
+1.000000 14.652657 1.386565 15.569288 1.921045 16.183510 c
+2.460337 16.803263 3.204824 17.178305 4.026860 17.292816 c
+5.686298 17.523981 7.572734 16.683088 8.418258 14.764525 c
+10.248410 15.571085 l
+9.019859 18.358770 6.239629 19.620373 3.750919 19.273689 c
+2.498881 19.099277 1.298923 18.515308 0.412288 17.496389 c
+-0.479158 16.471943 -1.000000 15.070246 -1.000000 13.357670 c
+1.000000 13.357670 l
+h
+10.248410 14.764524 m
+11.093934 16.683088 12.980371 17.523981 14.639809 17.292816 c
+15.461844 17.178305 16.206331 16.803263 16.745623 16.183510 c
+17.280104 15.569288 17.666668 14.652657 17.666668 13.357670 c
+19.666668 13.357670 l
+19.666668 15.070246 19.145826 16.471943 18.254379 17.496389 c
+17.367746 18.515308 16.167788 19.099277 14.915750 19.273689 c
+12.427039 19.620373 9.646809 18.358770 8.418258 15.571085 c
+10.248410 14.764524 l
+h
+17.666668 13.357670 m
+17.666668 12.115266 17.183661 10.865780 16.375643 9.650703 c
+15.568734 8.437297 14.474319 7.314231 13.348269 6.346621 c
+12.226038 5.382293 11.098922 4.594908 10.249951 4.047773 c
+9.826476 3.774857 9.474709 3.563320 9.230661 3.421019 c
+9.108696 3.349902 9.013809 3.296179 8.950520 3.260829 c
+8.918880 3.243156 8.895154 3.230086 8.879907 3.221739 c
+8.872284 3.217566 8.866782 3.214575 8.863473 3.212780 c
+8.861818 3.211883 8.860711 3.211286 8.860162 3.210989 c
+8.859887 3.210840 8.859752 3.210768 8.859756 3.210771 c
+8.859759 3.210772 8.859869 3.210831 8.859870 3.210831 c
+8.860014 3.210909 8.860194 3.211005 9.333334 2.330018 c
+9.806474 1.449030 9.806725 1.449165 9.807012 1.449320 c
+9.807155 1.449396 9.807477 1.449570 9.807763 1.449724 c
+9.808335 1.450033 9.809051 1.450417 9.809906 1.450878 c
+9.811618 1.451803 9.813893 1.453033 9.816725 1.454567 c
+9.822388 1.457638 9.830273 1.461927 9.840309 1.467422 c
+9.860381 1.478411 9.889064 1.494221 9.925784 1.514730 c
+9.999212 1.555744 10.104847 1.615585 10.238091 1.693277 c
+10.504459 1.848595 10.881859 2.075657 11.333384 2.366650 c
+12.234413 2.947336 13.440631 3.789028 14.651732 4.829722 c
+15.859015 5.867134 17.097935 7.125036 18.041027 8.543232 c
+18.983009 9.959758 19.666668 11.592510 19.666668 13.357670 c
+17.666668 13.357670 l
+h
+9.333334 2.330018 m
+9.806474 3.211005 9.806654 3.210909 9.806798 3.210831 c
+9.806799 3.210831 9.806909 3.210772 9.806911 3.210771 c
+9.806916 3.210768 9.806781 3.210840 9.806506 3.210989 c
+9.805957 3.211286 9.804850 3.211883 9.803195 3.212780 c
+9.799886 3.214575 9.794384 3.217566 9.786761 3.221739 c
+9.771514 3.230086 9.747787 3.243156 9.716148 3.260829 c
+9.652859 3.296179 9.557972 3.349902 9.436007 3.421019 c
+9.191959 3.563320 8.840192 3.774857 8.416717 4.047773 c
+7.567747 4.594908 6.440630 5.382293 5.318399 6.346621 c
+4.192348 7.314231 3.097934 8.437297 2.291026 9.650703 c
+1.483007 10.865780 1.000000 12.115266 1.000000 13.357670 c
+-1.000000 13.357670 l
+-1.000000 11.592510 -0.316341 9.959758 0.625641 8.543232 c
+1.568733 7.125036 2.807652 5.867134 4.014935 4.829722 c
+5.226037 3.789028 6.432254 2.947336 7.333284 2.366650 c
+7.784809 2.075657 8.162209 1.848595 8.428577 1.693277 c
+8.561821 1.615585 8.667456 1.555744 8.740884 1.514730 c
+8.777604 1.494221 8.806287 1.478411 8.826359 1.467422 c
+8.836395 1.461927 8.844280 1.457638 8.849943 1.454567 c
+8.852775 1.453033 8.855050 1.451803 8.856762 1.450878 c
+8.857617 1.450417 8.858333 1.450033 8.858905 1.449724 c
+8.859191 1.449570 8.859513 1.449396 8.859656 1.449320 c
+8.859943 1.449165 8.860194 1.449030 9.333334 2.330018 c
+h
+f
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  4231
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 24.000000 24.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Pages 5 0 R
+     /Type /Catalog
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000034 00000 n
+0000004321 00000 n
+0000004344 00000 n
+0000004517 00000 n
+0000004591 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+4650
+%%EOF
